requirements.txt
flask_cli
flask_sqlalchemy
psycopg2
dependency_injector
colorlog